Task 29.3 - TaskController changes to tests

+++ src/main/java/com/crud/tasks/controller/TaskContoller.java
@@ -11,36 +11,39 @@ import java.util.List;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 
 @RestController
-@RequestMapping("/v1/task")
+@RequestMapping("/v1")
 public class TaskContoller {
     @Autowired
     private DbService service;
     @Autowired
     private TaskMapper taskMapper;
 
-    @RequestMapping (method = RequestMethod.GET, value = "getTasks")
-    public List<TaskDto> getTasks(){
+    @RequestMapping(method = RequestMethod.GET, value = "/tasks")
+    public List<TaskDto> getTasks() {
         return taskMapper.mapToTaskDtoList(service.getAllTasks());
     }
 
-    @RequestMapping(method = RequestMethod.GET, value = "getTask")
-    public TaskDto getTasks(@RequestParam Long taskId) throws TaskNotFoundException{
-        return taskMapper.mapToTaskDto(service.getTaskById(taskId)
-        .orElseThrow(() -> new TaskNotFoundException("Task doesn't exist")));
+    @RequestMapping(method = RequestMethod.GET, value = "/tasks/{taskId}")
+    public TaskDto getTask(@PathVariable Long taskId) throws TaskNotFoundException {
+        return taskMapper.mapToTaskDto(service.getTaskById(taskId).orElseThrow(TaskNotFoundException::new));
     }
 
-    @RequestMapping(method = RequestMethod.POST, value = "createTask", consumes = APPLICATION_JSON_VALUE)
-    public void createTask(@RequestBody TaskDto taskDto){
-        service.saveTask(taskMapper.mapToTask(taskDto));
+    @RequestMapping(method = RequestMethod.DELETE, value = "/tasks/{taskId}")
+    public void deleteTask(@PathVariable Long taskId) throws TaskNotFoundException {
+        if (service.getTaskById(taskId).isPresent()) {
+            service.deleteTaskByID(taskId);
+        } else {
+            throw new TaskNotFoundException();
+        }
     }
 
-    @RequestMapping (method = RequestMethod.PUT, value = "updateTask")
-    public TaskDto updateTask(@RequestBody TaskDto taskDto){
+    @RequestMapping(method = RequestMethod.PUT, value = "/tasks")
+    public TaskDto updateTask(@RequestBody TaskDto taskDto) {
         return taskMapper.mapToTaskDto(service.saveTask(taskMapper.mapToTask(taskDto)));
     }
 
-    @RequestMapping(method = RequestMethod.DELETE, value = "deleteTask")
-    public void deleteTask(@RequestParam Long taskId) {
-        service.deleteTaskByID(taskId);
+    @RequestMapping(method = RequestMethod.POST, value = "/tasks", consumes = APPLICATION_JSON_VALUE)
+    public void createTask(@RequestBody TaskDto taskDto) {
+        service.saveTask(taskMapper.mapToTask(taskDto));
     }
 }
